Change rsync paths for heidrun-mappings
Change the rsync path arguments in the build script for heidrun, so that
heidrun-mappings gets rsynced into place without its top-level
directory's contents.  Only the `heidrun` directory needs to be
transferred, and the changes suggested by
https://github.com/dpla/heidrun-mappings/pull/28 will result in a `spec`
directory that will have files with `require` statements that will
result in load errors.